Formulario Ventas - Corrección errores - Manejos de Cta cte

+++ TPC_Barrachina/Negocio/Utilidades.cs
@@ -224,7 +224,7 @@ namespace Negocio
             throw new Exception("El " + NombreTextBox + " ingresado no existe");
         }
 
-        public decimal CalcularDescuento(int Subtotal,decimal Descuento) {
+        public decimal CalcularDescuento(decimal Subtotal,decimal Descuento) {
 
             return Subtotal - (Subtotal * Descuento / 100);
         }
